+++ client/context/types.go
@@ -22,6 +22,9 @@ type CoreContext struct {
 	Decoder         auth.AccountDecoder
 	AccountStore    string
 	UseLedger       bool
+	Async           bool
+	JSON            bool
+	PrintResponse   bool
 }
 
 // WithChainID - return a copy of the context with an updated chainID